+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.2.0
+ August 22, 2019
+- AppStateModel() in the View, ShrineApp
+- _BackdropState extends State
+- SetState(builder: (context, object) =>
+
 ## 0.1.0
  August 21, 2019
 - Initial commit